+++ Source/Gorgon/CGI/Line.h
@@ -36,7 +36,7 @@ namespace Gorgon :: CGI {
             do {
                 l = p.GetLine(s);
                 s++;
-            } while(s < p.GetSize() && l.Start == l.End);
+            } while(s < (int)p.GetSize() && l.Start == l.End);
             
             //nothing to draw here
             if(l.Start == l.End)
@@ -64,7 +64,7 @@ namespace Gorgon :: CGI {
 		}
 
         
-        for(int i=s; i<p.GetSize()-1; i++) {
+        for(int i=s; i<(int)p.GetSize()-1; i++) {
             Geometry::Line<P_> l = p.GetLine(i);
             
             if(l.Start == l.End)
@@ -136,7 +136,7 @@ namespace Gorgon :: CGI {
             do {
                 l = p.GetLine(st);
                 st++;
-            } while(st < p.GetSize() && l.Start == l.End);
+            } while(st < (int)p.GetSize() && l.Start == l.End);
 
 			auto off = (Geometry::Pointf(l.End) - Geometry::Pointf(l.Start)).Perpendicular().Normalize() * w;
