+++ Source/Gorgon/Resource/Writer.h
@@ -66,25 +66,25 @@ namespace Gorgon :: Resource {
 			Marker(const Marker &) = delete;
 			Marker(Marker &&other) {
 				pos = other.pos;
-				other.pos = -1;
+				other.pos = (unsigned long)-1;
 			}
 
 			Marker &operator =(Marker &&other) {
 				pos = other.pos;
-				other.pos = -1;
+				other.pos = (unsigned long)-1;
 
 				return *this;
 			}
 			
 			~Marker() { 
-				if(pos!=-1) 
+				if(pos!=(unsigned long)-1) 
 					Utils::ASSERT_FALSE("Marker is not ended.");
 			}
 			
 		private:
 			Marker(unsigned long pos) : pos(pos) { }
 			
-			unsigned long pos = -1;
+			unsigned long pos = (unsigned long)-1;
 		};
 		
 		/// Any writer implementation should close and set the stream to nullptr in destructor
